Clean up zone observation dupes, add validation
The zone name has been moved to a field on the resource, rather than
using the resource name, as it is an immutable field. Validation has
also been added to match the cloudflare API spec.

Some zone observation fields have been renamed to better communicate the
information the contain. Some fields duplicated between the observation
and the spec have been refactored out, with UpToDate and LateInitialize
using a cloudflare.Zone instance directly.

A ZoneID field has been added to the zone observation, which is used
instead of external-id. An ExtractZoneID function has been added to the
API which can be used in ResolveReferences calls that need the Zone ID.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>