Switch back to uninitialised external-name as uid
It turns out that the default behaviour of crossplane to initialise the
external-name annotation of a resource can be overridden in the Setup
method of the controller. We pass an empty Initializers{} setting, which
stops the external-name field from being pre-set, and allows us to
assume that if the field is not the zero value (empty string), then it
is a valid ID for a Cloudflare Zone.

This allows us to remove some additional code around sanity checking the
external-name and means we can use the standard ExternalName
ExtractValueFn to pull back the Zone UID.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>