Remove CreateFilters length check
If the Cloudflare API were returning multiple or zero Filters for a single
create call, we could trigger a panic by returning `nil,nil` from our
firewall client `CreateFilter` method (which we then attempt to
dereference).

We remove this extraneous length check, which means if at least one
Filter is returned from the CreateFilters method, then we'll proceed as
normal (if multiple Filters are created from 1 input then that's
Cloudflare's problem).

If the CreateFilters method does not error but DOES return 0 filters
then we can assume that something is very wrong with the Cloudflare API
and a panic is appropriate.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>