Re-add sanity check of CreateFilters retval
Thinking on this further, I'd rather we don't panic as much as possible,
as this potentially causes other resources on the provider to become 'hung'.

I've re-added a sanity check to the CreateFilter method that will return
an error if the underlying CreateFilters call does not return a single
Filter.

This may indicate the call created multiple filters, or none, both of
which are undefined behaviour from the API so logging this as a failed
creation attempt should be acceptable, while relying on Cloudflare's API
sanity check to not create duplicate filter expressions.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>