Fix ordering of UpToDate call
In some of our reconcilers we were ordering the call to our UpToDate
method after the LateInitialize call. I don't believe this change will
have any impact _now_, but it sets up a consistent pattern for the
future where it is possible that a field that has been LateInitialized
triggers the need for an update. If we do not run in this order, updates
may not be picked up until the next reconciliation loop.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>