Enable token auth
NOTE: This commit introduces a backwards incompatible change that
changes the JSON key of the existing "APIKey" and "Email" fields to
"apiKey" and "email". This was necessary as the field names did not
follow convention in the first place.

Cloudflare allows multiple authentication methods to the API - our
initial implementation was using an API key and the users' email. This
authentication is user-wide, and has no restrictions on the API.

To enable the provider to act in a non-superuser-manner, we now
implement authentication using API tokens, which can be created by a
user with reduced permissions.

Please note if the token you use does not have permissions to manage the
resources you want to deploy using crossplane, you will receive errors.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>