Add zone ignore fields
Zones were continually late initialised as the Cloudflare API returns
settings names that we do not (currently) support. In order to aid rapid
development, we will ignore these fields for the moment and implement
these at a later date.

We also fix the vanityNameServers late-init check here which was not
taking the remote state of vanityNS into account before late initing.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>