Rework Zone LateInit / UpToDate
Our previous `LateInitialize` and `UpToDate` implementations were rather
brittle, and meant changes on the Cloudflare side (sending fields we
were not expecting) would cause a `Zone` to be eternally late-initing or
updating.

We solve this by changing the following:

  - Not exporting `ZoneSettingsMap` outside of the `Zone` client. This is an
    implementation detail and passing it around alongside the
    `ZoneSettings` struct was confusing and redundant. From a controller
    perspective, we only use the `ZoneSettings` struct now. This means
    that any client function that needs to work on a `ZoneSettingsMap`
    must convert the `ZoneSettings` struct they are given first.
  - We make sure that `LoadSettingsForZone` always returns a `ZoneSettings`
    struct rather than an 'unfiltered' `ZoneSettingsMap`. This means that by
    the time we have access to the values, any unexpected information
    from the Cloudflare API has been removed. We only detect
    late-initialisation and whether the resource is up to date on settings
    that we know about. If Cloudflare add settings, we will not late-init
    or update the resource until we implement those fields in the client.
  - Settings comparison is now part of `UpToDate` rather than its' own
    `cmp.Equal` in the controller.

NOTE: We may still have issues with comparisons of slices, if those
slices are not correctly ordered. We can either solve this by always
ordering the values when retrieved from the API at either end, or use a
`cmpopts.SortSlices` function to handle this on comparison - but this
will be tackled in a separate MR.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>