Refactor custom hostname client change detection
This is a chunky refactor that avoids repeated use of if statements to
set and check whether a custom hostname is up to date. It uses a new
conversion method that allows us to build up a spec definition from the
Cloudflare API representation of a Custom Hostname for comparison with
our existing spec.

On Update or create, we use a new reverse method which turns a spec
definition into a Cloudflare API representation.

This also applies various fixes to the Custom Hostname API spec (missing
annotations mostly), and fixes some holes in the tests which only became
apparent with the new comparison method.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>